requirements.txt
# Playwright相关
playwright==1.40.0
pytest-playwright==0.4.3

# Pytest相关
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
pytest-timeout==2.2.0
pytest-json-report==1.5.0

# 数据处理
pyyaml==6.0.1
openpyxl==3.1.2
pandas==2.1.3

# 日志和报告
allure-pytest==2.13.2
loguru==0.7.2

# 工具类
python-dotenv==1.0.0
faker==20.1.0

